what do we get on test

+++ openpifpaf_detection_attributes/datasets/generators.py
@@ -241,7 +241,7 @@ class BoxGaussianAttributeGenerator(AttributeGenerator):
             gy = np.exp(-(y-y0)**2/(2*sigma_y**2))
             gx = np.exp(-(x-x0)**2/(2*sigma_x**2))
             g  = np.outer(gy, gx)
-            self.targets = np.nanmax([self.targets, np.expand_dims(g*obj[self.config.meta.attribute], axis=0).astype(np.float32)], axis=0).astype(np.float32)
+            self.targets = np.nanmax([self.targets, np.expand_dims(g*obj[self.config.meta.attribute], axis=0)], axis=0).astype(np.float32)
             print(obj["is_not_crossing_reg"], obj["is_crossing_reg"], obj[self.config.meta.attribute], self.config.meta.attribute)
             sys.stdout.flush()
